refactor(db): add MIN_RETENTION_DAYS guard to prediction_record cleanup
Align prediction_record::delete_old_records with the new cleanup pattern
used by other tables (config_store, evaluation_period, pool_info,
token_rate). Change retention_days from i64 to u32, add zero-check,
MIN_RETENTION_DAYS clamp, and structured logging.

+++ crates/common/src/config/typed.rs
@@ -557,13 +557,13 @@ define_typed_config! {
     // ── prediction ──
 
     /// Retention days for evaluated prediction records
-    fn prediction_record_retention_days() -> i64 {
+    fn prediction_record_retention_days() -> u32 {
         key: "PREDICTION_RECORD_RETENTION_DAYS",
         default: 30
     }
 
     /// Retention days for unevaluated predictions
-    fn prediction_unevaluated_retention_days() -> i64 {
+    fn prediction_unevaluated_retention_days() -> u32 {
         key: "PREDICTION_UNEVALUATED_RETENTION_DAYS",
         default: 20
     }
